# server/src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Todo App API Specification
  version: '69'
  description: API documentation for the Todo App bounded context.
servers:
  - url: /api/v1
    description: Development Server
...
paths:
...
  /update-todo:
    post:
      summary: Update Todo
      description: Executes the "Update Todo" command. ...
      operationId: updateTodo
      tags:
        - TodoApp
      responses:
        '200':
          description: 'OK: The resource was updated successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          $ref: '#/components/responses/BadRequest'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
...
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: An error occurred.
    Todo:
      type: object
      description: Represents a Todo in the system.
      properties:
        todoID:
          type: string
          description: The Todo ID of the Todo.
          example: '1'
        todoTitle:
          type: string
          description: The Todo Title of the Todo.
          example: Buy groceries
        description:
          type: string
          description: The Description of the Todo.
          example: Milk, eggs, bread
        dueDate:
          type: string
          description: The Due Date of the Todo.
          example: '2023-10-15'
        priority:
          type: string
          description: The Priority of the Todo.
          example: High
        category:
          type: string
          description: The Category of the Todo.
          example: Shopping
        completionDate:
          type: string
          description: The Completion Date of the Todo.
          example: '2023-10-16'
        completionNotes:
          type: string
          description: The Completion Notes of the Todo.
          example: Done
        deletionReason:
          type: string
          description: The Deletion Reason of the Todo.
          example: Not needed
      required:
        - todoID
        - todoTitle
...
    UpdateTodoRequest:
      type: object
      properties:
        todoID:
          type: string
          example: '1'
        todoTitle:
          type: string
          example: Buy groceries
        description:
          type: string
          example: Milk, eggs, bread
        dueDate:
          type: string
          example: '2023-10-15'
        priority:
          type: string
          example: High
        category:
          type: string
          example: Shopping
      required:
        - todoID
        - todoTitle
...
  responses:
    BadRequest:
      description: 'Bad Request: The server cannot or will not process the request due to something that is perceived to be a client error.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...